Added "home" library to handle file dialog locations
The "home" crate has been added to the dependencies in the Cargo.lock and Cargo.toml files. This library has been included to set the initial location for the file dialog in the window manager (src/window_manager/home.rs). This change will improve usability by starting the file dialog in the user's home directory.