Add support for file path to View struct
Added a new member `opened_path` to `View` struct that holds the path of currently opened file. Removed the `Default` implementation as it's no longer needed. In addition, made sure the list of previously opened files in `home.rs` does not have duplicates.

These changes allow to keep track of the specific file that the user has opened, improving the usability of the application by providing a more tailored user experience. Now, when switching views, the application will remember and display the path of the last file that was opened.