Update open crate version and refactor file viewer
Updated the 'open' crate version from 4.2.0 to 5.0.0 in Cargo.toml and Cargo.lock. Refactored the 'View' struct in 'view.rs' to include file reading functionality, and added a new enumerator 'ActiveTask' for current task status. Removed 'Default' for 'View' in 'mod.rs'.

This update is to improve file handling efficiency and to display the current task status on the user interface. Furthermore, updated crate 'open' to ensure compatibility and up-to-date dependency.